requirements.txt
paddleocr>=3.0.0
opencv-python>=4.10.0
Pillow>=10.4.0
numpy>=1.26.4
pydantic>=2.8.2
scikit-image>=0.24.0
shapely>=2.0.6
requests>=2.32.0
